refactor: refactoring Buttons components
