+++ src/views/Declaration/DeclarationForm.tsx
@@ -12,7 +12,7 @@ import MesuresCorrection from "../../components/MesuresCorrection"
 import { logToSentry, parseDate } from "../../utils/helpers"
 import RadiosBoolean from "../../components/RadiosBoolean"
 import { isFormValid, parseBooleanFormValue, parseBooleanStateValue, required } from "../../utils/formHelpers"
-import ButtonAction from "../../components/ButtonAction"
+import ButtonAction from "../../components/ds/ButtonAction"
 import ErrorMessage from "../../components/ErrorMessage"
 import { resendReceipt } from "../../utils/api"
 import FormStack from "../../components/ds/FormStack"
