+++ src/views/Indicateur1/IndicateurUnCoef/IndicateurUnCoefGroupForm.tsx
@@ -11,7 +11,7 @@ import globalStyles from "../../../utils/globalStyles"
 import { useColumnsWidth, useLayoutType } from "../../../components/GridContext"
 import InfoBlock from "../../../components/ds/InfoBlock"
 import ActionLink from "../../../components/ActionLink"
-import ButtonAction from "../../../components/ButtonAction"
+import ButtonAction from "../../../components/ds/ButtonAction"
 import ActionBar from "../../../components/ActionBar"
 import FormAutoSave from "../../../components/FormAutoSave"
 import FormSubmit from "../../../components/FormSubmit"
