+++ src/views/InformationsDeclarant/InformationsDeclarantForm.tsx
@@ -12,7 +12,7 @@ import FormAutoSave from "../../components/FormAutoSave"
 import FormSubmit from "../../components/FormSubmit"
 import FakeInputGroup from "../../components/ds/FakeInputGroup"
 import { ButtonSimulatorLink } from "../../components/SimulatorLink"
-import ButtonAction from "../../components/ButtonAction"
+import ButtonAction from "../../components/ds/ButtonAction"
 import { IconEdit } from "../../components/ds/Icons"
 import InputGroup from "../../components/ds/InputGroup"
 import FormStack from "../../components/ds/FormStack"
