+++ src/views/InformationsEntreprise/components/EntrepriseUESModalConfirmDelete.tsx
@@ -4,7 +4,7 @@ import { css, jsx } from "@emotion/react"
 import globalStyles from "../../../utils/globalStyles"
 
 import ActionLink from "../../../components/ActionLink"
-import ButtonAction from "../../../components/ButtonAction"
+import ButtonAction from "../../../components/ds/ButtonAction"
 
 import { IconWarning } from "../../../components/ds/Icons"
 import { useColumnsWidth, useLayoutType } from "../../../components/GridContext"
