+++ src/views/InformationsSimulation/InformationsSimulationForm.tsx
@@ -19,7 +19,7 @@ import InputDateGroup from "../../components/ds/InputDateGroup"
 import FormAutoSave from "../../components/FormAutoSave"
 import FormSubmit from "../../components/FormSubmit"
 import { ButtonSimulatorLink } from "../../components/SimulatorLink"
-import ButtonAction from "../../components/ButtonAction"
+import ButtonAction from "../../components/ds/ButtonAction"
 import FormStack from "../../components/ds/FormStack"
 import { IconEdit } from "../../components/ds/Icons"
 import InputRadioGroup from "../../components/ds/InputRadioGroup"
